Move v4 API to its own package
